Revert "using for readv/writev the same interface like UNIX"
This reverts commit bf0cdbd9cbaaad4bf118cb8200c4a3c7da45b949.

+++ hermit-abi/src/lib.rs
@@ -140,7 +140,7 @@ pub const POLLRDHUP: i16 = 0x2000;
 pub const EFD_SEMAPHORE: i16 = 0o1;
 pub const EFD_NONBLOCK: i16 = 0o4000;
 pub const EFD_CLOEXEC: i16 = 0o40000;
-pub const IOV_MAX: i32 = 1024;
+pub const IOV_MAX: usize = 1024;
 pub type sa_family_t = u8;
 pub type socklen_t = u32;
 pub type in_addr_t = u32;
@@ -622,7 +622,7 @@ extern "C" {
 	/// which data should be written.  `readv()` will always fill an completely
 	/// before proceeding to the next.
 	#[link_name = "sys_readv"]
-	pub fn readv(fd: i32, iov: *const iovec, iovcnt: i32) -> isize;
+	pub fn readv(fd: i32, iov: *const iovec, iovcnt: usize) -> isize;
 
 	/// `getdents64` reads directory entries from the directory referenced
 	/// by the file descriptor `fd` into the buffer pointed to by `buf`.
@@ -681,7 +681,7 @@ extern "C" {
 	/// which data should be written.  `writev()` will always write a
 	/// complete area before proceeding to the next.
 	#[link_name = "sys_writev"]
-	pub fn writev(fd: i32, iov: *const iovec, iovcnt: i32) -> isize;
+	pub fn writev(fd: i32, iov: *const iovec, iovcnt: usize) -> isize;
 
 	/// close a file descriptor
 	///
